+++ Source/ModelUtils/Colliders.cs
@@ -140,6 +140,8 @@ public static class Colliders {
   /// <see cref="PrimitiveType.Cylinder"/>, <see cref="PrimitiveType.Sphere"/>, and
   /// <see cref="PrimitiveType.Cube"/> are supported.
   /// </param>
+  /// FIXME: it's not working with asymetric meshes
+  /// FIXME: use GameOvbject.GetRedererBounds?
   public static void AdjustCollider(
       GameObject primitive, Vector3 meshSize, PrimitiveCollider colliderType,
       PrimitiveType? shapeType = null) {
@@ -222,6 +224,61 @@ public static class Colliders {
     var modelRoot = Hierarchy.GetPartModelTransform(part);
     vessel.parts.ForEach(
         p => SetCollisionIgnores(modelRoot, Hierarchy.GetPartModelTransform(p), ignore));
+  }
+
+  /// <summary>
+  /// Returns the minimum square distance to the nearest point on the part's collider surface.
+  /// </summary>
+  /// <remarks>This method skips triggers and inactive colliders.</remarks>
+  /// <param name="point">The reference point to find distance for.</param>
+  /// <param name="part">The part to check for.</param>
+  /// <param name="filterFn">
+  /// The filter function to apply to every collider. Return <c>false</c> from it to sklip the
+  /// collider in the following checks.
+  /// </param>
+  /// <returns>The square distance or <c>null</c> if no colliders found.</returns>
+  public static float? GetSqrDistanceToPart(
+      Vector3 point, Part part, Func<Collider, bool> filterFn = null) {
+    float? minDistance = null;
+    var colliders = part.transform.GetComponentsInChildren<Collider>()
+        .Where(c => !c.isTrigger && c.enabled && c.gameObject.activeInHierarchy
+               && (filterFn == null || filterFn(c)));
+    foreach (var collider in colliders) {
+      Vector3 closetsPoint;
+      if (collider is WheelCollider) {
+        // Wheel colliders don't support closets point check.
+        closetsPoint = collider.ClosestPointOnBounds(point);
+      } else {
+        closetsPoint = collider.ClosestPoint(point);
+        if (closetsPoint == collider.transform.position) {
+          // The point it inside the collider or on the boundary.
+          return 0.0f;
+        }
+      }
+      var sqrMagnitude = (closetsPoint - point).sqrMagnitude;
+      minDistance = Mathf.Min(minDistance ?? float.PositiveInfinity, sqrMagnitude);
+    }
+    return minDistance;
+  }
+
+  /// <summary>
+  /// Returns the minimum square distance to the nearest point on the part's collider surface.
+  /// </summary>
+  /// <remarks>This method skips triggers and inactive colliders.</remarks>
+  /// <param name="point">The reference point to find distance for.</param>
+  /// <param name="part">The part to check for.</param>
+  /// <param name="defaultValue">The value to return if no suitable colliders found.</param>
+  /// <param name="filterFn">
+  /// The filter function to apply to every collider. Return <c>false</c> from it to sklip the
+  /// collider in the following checks.
+  /// </param>
+  /// <returns>
+  /// The square distance or <paramref name="defaultValue"/> if no colliders found.
+  /// </returns>
+  public static float GetSqrDistanceToPartOrDefault(
+      Vector3 point, Part part,
+      float defaultValue = float.PositiveInfinity, Func<Collider, bool> filterFn = null) {
+    return GetSqrDistanceToPart(point, part) ?? defaultValue;
   }
 }